+++ gradio/__init__.py
@@ -1,8 +1,8 @@
 import pkg_resources
 
-from gradio.app import get_state, set_state
+from gradio.routes import get_state, set_state
 from gradio.flagging import *
-from gradio.interface import *  # This makes it possible to import `Interface` as `gradio.Interface`.
+from gradio.interface import * 
 from gradio.mix import *
 
 current_pkg_version = pkg_resources.require("gradio")[0].version
